move RenderAPI to its own file
